requirements.txt
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
click==8.1.3
confection==0.0.4
cymem==2.0.7
Flask==2.3.1
itsdangerous==2.1.2
langcodes==3.3.0
murmurhash==1.0.9
numpy==1.24.3
pathy==0.10.1
pdfminer.six==20221105
preshed==3.0.8
pydantic==1.10.7
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
wasabi==1.1.1
Werkzeug==2.3.1
